refactor: 블로그 헤더 슬롯 구조 통합
