--- src/app/(blog-layout)/@head-section/p/[title]/page.tsx
@@ -1,26 +1,0 @@
-import { ContentHeader } from "@/components/head-section/content-header";
-import { getPostFromParamsBySlug, type PostParams } from "@/utils/post-util";
-
-interface IPostHeadSectionProps {
-  params: Promise<PostParams>;
-}
-
-const PostHeadSection = async ({ params }: IPostHeadSectionProps) => {
-  const post = getPostFromParamsBySlug(await params);
-
-  if (!post) {
-    return null;
-  }
-
-  return (
-    <ContentHeader
-      title={post.title}
-      text={post.description}
-      img={post.thumbnail}
-      tags={post.tags}
-      date={post.date}
-    />
-  );
-};
-
-export default PostHeadSection;
