--- src/app/(blog-layout)/@head-section/page.tsx
@@ -1,13 +1,0 @@
-import { ContentHeader } from "@/components/head-section/content-header";
-
-const HeadSection = () => {
-  return (
-    <ContentHeader
-      title="Insights"
-      text="노력과 고민을 담아 기록한 글"
-      main
-    />
-  );
-};
-
-export default HeadSection;
