--- src/app/(blog-layout)/@right-section/p/[title]/page.tsx
@@ -1,11 +1,0 @@
-import { TOCList } from "@/components/right-section/toc-iist";
-
-const PostRightSection = () => {
-  return (
-    <div className="w-full h-auto sticky top-32 list-none ">
-      <TOCList />
-    </div>
-  );
-};
-
-export default PostRightSection;
