--- src/app/(blog-layout)/@right-section/page.tsx
@@ -1,7 +1,0 @@
-import { TagsBox } from "@/components/right-section/tags-box";
-
-const RightSection = () => {
-  return <TagsBox />;
-};
-
-export default RightSection;
