--- src/app/(blog-layout)/@right-section/tags/[tag]/page.tsx
@@ -1,7 +1,0 @@
-import { TagsBox } from "@/components/right-section/tags-box";
-
-const TagsRightSection = () => {
-  return <TagsBox />;
-};
-
-export default TagsRightSection;
