+++ src/app/(blog-layout)/layout.tsx
@@ -2,17 +2,11 @@ import { ReactNode } from "react";
 
 interface BlogLayoutProps {
   children: ReactNode;
-  "head-section": ReactNode;
-  "right-section": ReactNode;
 }
 
-const BlogLayout = ({
-  children,
-  "head-section": headSection,
-}: BlogLayoutProps) => {
+const BlogLayout = ({ children }: BlogLayoutProps) => {
   return (
     <div className="w-dvw h-auto relative overflow-x-hidden">
-      {headSection}
       <main className="w-full">{children}</main>
     </div>
   );
