+++ src/components/content-header/image-with-placeholder-header.tsx
