+++ src/components/nav/site-header.tsx
@@ -3,8 +3,8 @@ import { PopoverButton } from "./header/popover-button";
 
 const Header = () => {
   return (
-    <header className="z-40 flex h-14 w-full items-center border-b border-slate-800 bg-dark-bg px-6 md:h-20">
-      <nav className="flex h-full w-full items-center justify-between">
+    <header className="z-40 flex h-14 w-full items-center border-b border-slate-800 bg-dark-bg px-6 md:h-[100px]">
+      <nav className="mx-auto flex h-full w-full max-w-[1632px] items-center justify-between">
         <Link href={"/"}>
           <div className="flex justify-center items-center font-bold text-2xl md:text-3xl cursor-pointer hover:group group">
             <p className="text-gradient-start">&lt;</p>
