--- src/components/right-section/tags-box.tsx
@@ -1,37 +1,0 @@
-import { getTagsFromPosts } from "@/utils/post-util";
-import Link from "next/link";
-
-interface TagItemProps {
-  title: string;
-  count: number;
-}
-
-const TagItem = ({ title, count }: TagItemProps) => {
-  return (
-    <Link href={`/tags/${title}`}>
-      <div className="w-full flex rounded-md py-1 px-3 items-center">
-        <p className="cursor-pointer text-gradient-start hover:underline">
-          {title.toUpperCase()}
-        </p>
-        <p className="text-slate-400 text-xs">&nbsp;({count})</p>
-      </div>
-    </Link>
-  );
-};
-
-const TagsBox = async () => {
-  const tagsCount = await getTagsFromPosts();
-
-  return (
-    <div className="shadow-md w-full h-auto p-3 sticky top-32 rounded-lg bg-content-header-black">
-      <p>Tags</p>
-      <div className="mt-8 flex flex-col space-y-3">
-        {tagsCount.map(({ tag, count }) => (
-          <TagItem key={"TAGITEM" + tag} title={tag} count={count} />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export { TagsBox };
